requirements.txt
# ============================================================
# DermAssist - Python Dependencies
# ============================================================

# --- Deep Learning ---
torch>=2.0.0
torchvision>=0.15.0

# --- API Framework ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# --- Image Processing ---
Pillow>=10.0.0

# --- Configuration ---
PyYAML>=6.0

# --- Utilities ---
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
